Citadel/services/orchestrator/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./services/orchestrator/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
async-timeout==5.0.1
    # via -r services/orchestrator/requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r services/orchestrator/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r services/orchestrator/requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
neo4j==5.28.2
    # via -r services/orchestrator/requirements.in
psycopg2-binary==2.9.10
    # via -r services/orchestrator/requirements.in
pydantic==2.11.7
    # via
    #   -r services/orchestrator/requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pytz==2025.2
    # via neo4j
redis==6.4.0
    # via -r services/orchestrator/requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via -r services/orchestrator/requirements.in
